README.md - include setup info for a server running `storm-alerts`
We use the storm-docker repository to run the storm-alerts topology.
Since storm-docker setup was done earlier today on a new machine and my
memory is still fresh, the README.md file is updated with instructions
on setting it up on a server which will run the storm-alerts topology.

Hopefully this will help someone else who needs to setup a new server
for storm-alerts in the future.

Signed-off-by: Pang Yan Han <yanhan@viki.com>