destroy-storm.sh - use `docker stop` instead of `docker kill`
Because `docker stop` does a SIGTERM before a SIGKILL to the main
process running inside the Docker container, which is the standard (and
more graceful) way to kill a UNIX process.

In addition, based on the `docker kill` documentation:

    http://docs.docker.io/reference/commandline/cli/#kill

there may be bugs with `docker kill` which could make it difficult to
remove the container.

Signed-off-by: Pang Yan Han <yanhan@viki.com>