Zookeeper - use custom zoo.cfg
A summary of the changes in this patch:
- set custom `dataDir`
- set custom `dataLogDir`
- auto purge snapshots
- set `ZOO_DATADIR_AUTOCREATE_DISABLE` environment variable

The `dataDir` is the directory where Zookeeper stores snapshots. Our
previous `zoo.cfg`, which is a copy of `zoo_sample.cfg`, uses
`/tmp/zookeper` as the `dataDir`, which is not recommended due to the
temporal nature of the `/tmp` directory. In this patch, we change
`dataDir` to `/var/lib/zookeeper` .

We point `dataLogDir` to the `/zookeeper-transactions` directory. This
is the least we could do as a workaround to placing `dataLogDir` on a
separate device, as highlighted in the `dataDir` and `dataLogDir`
sections in the Zookeeper documentation here:

    http://zookeeper.apache.org/doc/trunk/zookeeperAdmin.html#sc_configuration

Since Zookeeper does not automatically remove old snapshots and log
files (seen here:
http://zookeeper.apache.org/doc/trunk/zookeeperAdmin.html#Ongoing+Data+Directory+Cleanup),
this is done using the `autopurge.snapRetainCount` and
`autopurge.purgeInterval` configuration parameters in `zoo.cfg`.

The `ZOO_DATADIR_AUTOCREATE_DISABLE` environment variable is set to 1 to
prevent Zookeeper from creating a non-existent data directory in case
the `dataDir` configuration parameter is changed. This is recommended
for a production Zookeeper setup as stated here:

    http://zookeeper.apache.org/doc/trunk/zookeeperAdmin.html#Disabling+data+directory+autocreation

Signed-off-by: Pang Yan Han <yanhan@viki.com>