README.md - remove lxc related configuration
Previously, we were using lxc for convenience, since an `lxc-attach`
allows us to access a running Docker container without running an SSH
server.

From:

    http://www.infoq.com/news/2014/03/docker_0_9

Docker has dropped lxc as the default execution driver. In view of this,
we've decided to not use lxc as the default execution driver as well.

Signed-off-by: Pang Yan Han <yanhan@viki.com>