README.md - use `bash` instead of `sh` to run the script from get.docker.io
Some Linux systems alias /bin/sh to the dash shell and not the Bash
shell. If that's the case , the script from http://get.docker.io cannot run
correctly.

Using `bash` has worked so far, so we might as well convert use `bash`
in the Docker installation command so the user can smoothly install Docker.

Signed-off-by: Pang Yan Han <yanhan@viki.com>