Makefile - copy `config/storm-setup.yaml` to `base-storm/storm-setup.yaml` if different
Previously, we copy the `config/storm-setup.yaml` to
`base-storm/storm-setup.yaml` unconditionally. This caused Docker to not
use the cache for steps after `ADDing` the `storm-setup.yaml` file even
if the contents of the new file are the same as the old one. My guess is
that Docker is probably using timestamps of a file to be `ADDed` to
determine if it should perform the step or use the cache.

This patch copies the `config/storm-setup.yaml` file to
`base-storm/storm-setup.yaml` if their contents differ. We check for
difference by comparing the MD5 checksums of the source and destination
`storm-setup.yaml` file. By doing so, the `base-storm/storm-setup.yaml`
file is only copied on content change and this allows Docker to make use
of its much faster cache when appropriate.

Signed-off-by: Pang Yan Han <yanhan@viki.com>