[BUGFIX] Servers may not be from EC2, dont take non-zero return code as fatal
Because we've been testing on EC2 servers all the while, we've neglected
the fact that users not be using EC2 servers for running Storm. Instead
of raising a ValueError on a non zero return code from curl (used to
obtain the public and private IP addresses for an Amazon EC2 instance),
we instruct the `ec2_get_ip` function to return `None`.

Signed-off-by: Pang Yan Han <yanhan@viki.com>