Separate nimbus and drpc ports
Because we run the Storm nimbus and drpc in the Storm nimbus Docker,
the sections specifying ports in `storm.yaml` for these two components
are specified in the same list in the the `STORM_COMPONENTS_PORTS`
global (in `docker_python_helpers/docker_run.py`). This assumption of
running both nimbus and drpc in the same Docker may not be valid with
future modifications.

In this patch, we separate out the ports for these two components. This
makes it possible for someone who modifies the code in the future to run
nimbus and drpc in separate Docker containers.

Signed-off-by: Pang Yan Han <yanhan@viki.com>