[BUGFIX] Replace Zookeeper IP address for multi server setup at correct index
Previously, in the `run-supervisord.py` script, if there was a need to
replace the IP address of a Zookeeper server, the original IP address is
removed and the replacement IP address is appended to the list of
Zookeeper servers. This did not cause problems initially, but later on,
new Storm setups were not able to run.

This patch replaces the IP address at the appropriate index
corresponding to the id of the Zookeeper server.

Signed-off-by: Pang Yan Han <yanhan@viki.com>